refactor: CommentContent, CommentDateTime 업데이트
