+++ src/components/ArticleComment/CommentContent.tsx
@@ -1,26 +1,31 @@
 import { css } from '@emotion/react';
 
-import { fontCss } from '~/styles/utils';
+import { fontCss, palettes } from '~/styles/utils';
 
 interface CommentContentProps {
   content: string;
+  modified: boolean;
   className?: string;
 }
 
 const CommentContent = (props: CommentContentProps) => {
-  const { content, ...restProps } = props;
+  const { content, modified, className } = props;
 
   return (
-    <div css={selfCss} {...restProps}>
-      {content.split('\n').map((p, index) => {
-        return <CommentParagraph key={index} paragraph={p} />;
-      })}
+    <div className={className}>
+      <div css={selfCss}>
+        {content.split('\n').map((p, index) => {
+          return <CommentParagraph key={index} paragraph={p} />;
+        })}
+      </div>
+      {modified && <span css={modifiedCss}>(수정됨)</span>}
     </div>
   );
 };
 
 export default CommentContent;
 const selfCss = css(fontCss.style.R14);
+const modifiedCss = css({ color: palettes.primary.default }, fontCss.style.R14);
 
 interface CommentParagraphProps {
   paragraph: string;
